Add files via upload
The Occam Razor Principle. If a phenomenon can be explained in two ways: first through involvement of entities A, B, C or second through entities A, B, C and D, and both methods give same result, first explanation should be better. Essence of D in this example is superfluous and its involvement redundant. Code #ralf1 does same thing, it gets rid of parameters that may affect result but at same time they are not crucial. In fact, this is basic principle of logic and everything is logical in sense that we have accepted it in our universal understanding - if there is something extra we do not consider it logically justified, since we ask ourselves in this case, for what is it? This is done by creating matrix based on rearranging elements of vector using random number generator. This is analogous to wave we pass through code and normalize result so that old data have amplitude composition of spectrum as result of prediction. Step by step. That is, chaos is created and code determines where chaos is systematic. Even if you add some unknown points to input, then code will change those points as it would be logical based on system created from this data. code not only thinks, but also can not disturb logical course of things. Whatever points you put in. https://www.linkedin.com/posts/vadimchornyy_ralf1-activity-6646829973653073920-Ng1Y